Use object resolver
