+++ src/index.js
@@ -1,3 +1,2 @@
-export createElements from './createElements';
-export createResolveElements from './createResolveElements';
 export RelayRouteRenderer from './RelayRouteRenderer';
+export Resolver from './Resolver';
